+++ Kentico.MVC.Widgets/Raybiztech.Kentico12.MVC.Widgets.GoogleMap/packages.config
@@ -39,9 +39,9 @@
   <package id="Microsoft.CodeAnalysis.VisualBasic.Workspaces" version="3.3.1" targetFramework="net461" />
   <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="3.3.1" targetFramework="net461" />
   <package id="Microsoft.CSharp" version="4.4.1" targetFramework="net461" />
-  <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net461" />
+  <package id="Microsoft.Data.Edm" version="5.8.4" targetFramework="net461" />
   <package id="Microsoft.Data.OData" version="5.8.4" targetFramework="net461" />
-  <package id="Microsoft.Data.Services.Client" version="5.6.4" targetFramework="net461" />
+  <package id="Microsoft.Data.Services.Client" version="5.8.4" targetFramework="net461" />
   <package id="Microsoft.Owin" version="3.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security" version="3.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.Cookies" version="3.1.0" targetFramework="net461" />
@@ -82,7 +82,7 @@
   <package id="System.Reflection.Metadata" version="1.6.0" targetFramework="net461" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net461" />
   <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="net461" />
-  <package id="System.Spatial" version="5.6.4" targetFramework="net461" />
+  <package id="System.Spatial" version="5.8.4" targetFramework="net461" />
   <package id="System.Text.Encoding.CodePages" version="4.5.1" targetFramework="net461" />
   <package id="System.Threading.Tasks.Extensions" version="4.5.3" targetFramework="net461" />
   <package id="System.ValueTuple" version="4.4.0" targetFramework="net461" />
